Make color choice more obvious
Since the color can change based on the brightness, the color picked doesn't match what the light gets set to.

This causes the picked color to not be highlighted. So set the background of the color tab to the color

If we put lighter colors in there, we might need something to change the text color to black.